集成Mybatis官方代码生成器, 新增generator模块 1. 当然也可以使用 Mybatis Puls 封装Mybatis 2. Mybatis+官方代码生成器, 没有痛点, 只是再多用一层框架而已
官方代码生成器: Mybatis Generator, 生成器支持独立运行, 用法如下:
1. 下载jar包单独运行,生成文件手动拷贝
2. 集成到项目中来

集成到项目中需要引入mybatis官方代码插件
开发规范: 生成器生成的代码(这里是4个文件)不能手动修改

四个文件包括:
1. Member实体类
2. MemberExample
3. MemberMapper.java
4. MemberMapper.xml

自定义SQL需要写到自己的mapper里, 不能放在生成的mapper里, 因为会被覆盖, 同时生成器只能生成单表增删改查

代码生成器模版generator-config-member.xml

同时在本次过程当中出现了一个小bug,有一个"WARN  o.s.c.LocalVariableTableParameterNameDiscoverer:123"的警告, 可能mybatis的版本太低,但是实际更新之后依旧无法解决, 后期再研究, 不影响代码正常运行

+++ train-member/src/main/java/com/xiaoweii/train/member/controller/MemberController.java
@@ -17,7 +17,7 @@ public class MemberController {
     private MemberService memberService;
 
     @GetMapping("/count")
-    public int hello(){
+    public long hello(){
         return memberService.count();
     }
 }
